code refactor adding open search
