+++ cmd/services/webcrawler/internal/dynamoDBx/add.go
@@ -5,7 +5,7 @@ import (
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/feature/dynamodb/attributevalue"
 	"github.com/aws/aws-sdk-go-v2/service/dynamodb"
-	"webcrawler/site"
+	"webcrawler/pkg/site"
 )
 
 func (db *DB) AddPage(ctx context.Context, page site.Page) error {
