+++ cmd/services/webcrawler/internal/dynamoDBx/type.go
@@ -4,7 +4,7 @@ import (
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/dynamodb"
 	"os"
-	"webcrawler/awsx"
+	"webcrawler/pkg/awsx"
 )
 
 type DB struct {
