+++ cmd/services/webcrawler/internal/formating/text_test.go
@@ -3,7 +3,7 @@ package formating
 import (
 	"github.com/stretchr/testify/require"
 	"testing"
-	"webcrawler/site"
+	"webcrawler/pkg/site"
 )
 
 func TestFetchLinks(t *testing.T) {
