+++ cmd/services/webcrawler/internal/graphx/add.go
@@ -3,7 +3,7 @@ package graphx
 import (
 	"context"
 	"github.com/neo4j/neo4j-go-driver/v5/neo4j"
-	"webcrawler/site"
+	"webcrawler/pkg/site"
 )
 
 func (g *Graph) AddWebsite(ctx context.Context, web site.Website) error {
