+++ pkg/queue/remove.go
